refactor(downloads): #176 Addresses PR comments
