--- src/Components/PaperEditor/ProgrammeScriptContainer/ExportDropdown/ExportModal/index.module.css
@@ -1,13 +1,0 @@
-.mediaItem {
-    list-style-type: none;
-    margin: .85rem 0;
-}
-.status {
-    text-align: right;
-    font-size: .85rem;
-    margin-left: .85rem;
-}
-.mediaName {
-    margin-left: .65rem;
-    display: inline-block;
-}
